pyproject.toml
[project]
name = "filebox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "email-validator>=2.0.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.22",
    "psycopg2-binary>=2.9.11",
    "voyageai>=0.3.7",
    "pgvector>=0.3.0",
    "groq>=1.1.1",
]
